small fixes to variable declarations to get it compiling

+++ src/int_recom/recom_modules.F90
@@ -445,9 +445,9 @@ module recom_config
                           rho_ref_water, visc_ref_water, w_ref1, w_ref2, depth_scaling1,   &
                           depth_scaling2, max_sinking_velocity
 #if defined (__seaice_tracers)
-  REAL(kind=WP)  :: dust_sol_ice = 1.0            ! solubility of iron fro dust when moved through ice
-  REAL(kind=WP)  :: sitr_frac_from_freezing = 1.0 ! fraction of iron concentration built into fresh seaice when formed by freezing
-  REAL(kind=WP)  :: sitr_frac_from_flooding = 1.0 ! fraction of iron concentration built into fresh seaice when formed by flooding
+  REAL(kind=8)  :: dust_sol_ice = 1.0            ! solubility of iron fro dust when moved through ice
+  REAL(kind=8)  :: sitr_frac_from_freezing = 1.0 ! fraction of iron concentration built into fresh seaice when formed by freezing
+  REAL(kind=8)  :: sitr_frac_from_flooding = 1.0 ! fraction of iron concentration built into fresh seaice when formed by flooding
   namelist /pa_seaiceiron/ dust_sol_ice, sitr_frac_from_freezing, sitr_frac_from_flooding
 #endif /* (__seaice_tracers) */
